Minor fixes to logging, and code adjustments
