+++ src/main/resources/data/minestuck/minestuck/captcha_codes.json
